Warning cleanup.

+++ src/gui/dialogs/freedv_reporter.cpp
@@ -1005,7 +1005,7 @@ void FreeDVReporterDialog::OnStatusTextChange(wxCommandEvent& event)
     // reduces the number of times one must backspace in order to completely
     // remove the flag.
     int index = 0;
-    while (index < statusMsg.Length())
+    while (index < (int)statusMsg.Length())
     {
         auto chr = statusMsg.GetChar(index);
         if (chr.GetValue() == 0x1F3F4)
@@ -1013,7 +1013,7 @@ void FreeDVReporterDialog::OnStatusTextChange(wxCommandEvent& event)
             log_debug("Found start char at index %d", index);
             auto endIndex = index + 1;
             bool foundEnd = false;
-            while (endIndex < statusMsg.Length())
+            while (endIndex < (int)statusMsg.Length())
             {
                 auto endChar = statusMsg.GetChar(endIndex).GetValue();
                 if (endChar == 0xE007F)
